+++ dist-extensions/index.d.ts
@@ -1,2 +1,0 @@
-export * from './fabric';
-//# sourceMappingURL=index.d.ts.map