+++ extensions/index.ts
@@ -1,34 +1,5 @@
-export { AligningGuidelines } from '../packages/aligning-guidelines/src';
-export type * from '../packages/aligning-guidelines/src/typedefs';
-
-export {
-  originUpdaterWrapper,
-  installOriginWrapperUpdater,
-} from '../packages/data-updaters/src/origins';
-
-export {
-  gradientUpdaterWrapper,
-  installGradientUpdater,
-} from '../packages/data-updaters/src/gradient';
-
-export {
-  addGestures,
-  pinchEventHandler,
-  rotateEventHandler,
-} from '../packages/westures-integration/src';
-
-export {
-  createImageCroppingControls,
-  createImageResizeControlsWithScaleToCover,
-} from '../packages/cropping-controls/src/croppingControls';
-export {
-  changeCropY,
-  changeCropX,
-  changeCropWidth,
-  changeCropHeight,
-  changeWidthAndScaleToCover,
-  changeHeightAndScaleToCover,
-  withFlip,
-} from '../packages/cropping-controls/src/croppingHandlers';
-export { enterCropMode } from '../packages/cropping-controls/src/enterCropMode';
-export { createLinearGradientControls } from './linear_gradient_controls/linearGradientControls';
+export * from '@fabricjs/aligning-guidelines';
+export * from '@fabricjs/cropping-controls';
+export * from '@fabricjs/data-updaters';
+export * from '@fabricjs/gradient-controls';
+export * from '@fabricjs/westures-integration';