+++ index.ts
@@ -1,6 +1,1 @@
-import { setEnvFactory } from './packages/core/src/env';
-import { getEnv } from './packages/browser/src/env';
-
-setEnvFactory(getEnv);
-
-export * from './fabric';
+export * from '@fabricjs/browser';